requirements.txt
tinytag==2.1.0
Pillow>=11.0.0
black>=24.2.0
pymediainfo>=6.0.1
pypdf>=5.3.0  # Optional: for PDF metadata extraction
TKinterModernThemes
# Note: Tkinter should be installed system-wide using: brew install python-tk@3.13 

# Development dependencies
vulture>=2.14  # For finding unused code 